feat: split sseReq to another file
